Additional tweaks for Visor and more
Tweak the APIs around title pane control buttons - tracked by https://github.com/kirill-grouchnikov/substance/issues/42
Tweak early return in title pane layout for https://github.com/kirill-grouchnikov/substance/issues/46
Move VisorMail to substance-samples
Switch UiUtils.isRetina to more fine-level UiUtils.getScaleFactor
Expose getScaleFactor and getBlankImage as GlobalScope APIs

+++ src/org/pushingpixels/substance/api/SubstanceLookAndFeel.java
@@ -408,9 +408,9 @@ public abstract class SubstanceLookAndFeel extends BasicLookAndFeel {
                     icon.getIconHeight());
             Graphics2D g2d = intermediate.createGraphics();
             g2d.setComposite(AlphaComposite.SrcOver.derive(alpha));
-            int scaleFactor = UIUtil.getScaleFactor();
-            g2d.drawImage(result, 0, 0, result.getWidth() / scaleFactor,
-                    result.getHeight() / scaleFactor, null);
+            double scaleFactor = UIUtil.getScaleFactor();
+            g2d.drawImage(result, 0, 0, (int) (result.getWidth() / scaleFactor),
+                    (int) (result.getHeight() / scaleFactor), null);
             g2d.dispose();
             result = intermediate;
         }
