+++ src/org/pushingpixels/substance/api/painter/decoration/ClassicDecorationPainter.java
@@ -95,9 +95,9 @@ public class ClassicDecorationPainter implements SubstanceDecorationPainter {
 						width, height, scheme);
 				smallImageCache.put(key, result);
 			}
-			int imageScale = UIUtil.getScaleFactor();
-			graphics.drawImage(result, 0, 0, result.getWidth() / imageScale, 
-			        result.getHeight() / imageScale, null);
+			double imageScale = UIUtil.getScaleFactor();
+			graphics.drawImage(result, 0, 0, (int) (result.getWidth() / imageScale), 
+			        (int) (result.getHeight() / imageScale), null);
 			return;
 		}
