+++ src/org/pushingpixels/substance/internal/utils/border/SubstanceToolBarBorder.java
@@ -70,7 +70,7 @@ public class SubstanceToolBarBorder extends AbstractBorder implements
 
 		Graphics2D graphics = (Graphics2D) g.create();
 		graphics.translate(x, y);
-		int scaleFactor = UIUtil.getScaleFactor();
+		double scaleFactor = UIUtil.getScaleFactor();
 		graphics.scale(1.0f / scaleFactor, 1.0f / scaleFactor);
 
 		if (((JToolBar) c).isFloatable()) {
